+++ nonebot_plugin_saa/adapters/__init__.py
@@ -1,5 +1,6 @@
 from . import qq as qq
 from . import red as red
+from . import dodo as dodo
 from . import feishu as feishu
 from . import qqguild as qqguild
 from . import kaiheila as kaiheila