+++ nonebot_plugin_saa/utils/const.py
@@ -9,6 +9,7 @@ class SupportedAdapters(StrEnum):
     telegram = "Telegram"
     feishu = "Feishu"
     red = "RedProtocol"
+    dodo = "DoDo"
     qq = "QQ"
 
     fake = "fake"  # for nonebug
@@ -18,11 +19,10 @@ class SupportedPlatform(StrEnum):
     qq_group = "QQ Group"
     qq_group_open = "QQ Group Open"
     qq_private = "QQ Private"
-    qq_private_open = "QQ Private Open"
+    qq_group_openid = "QQ Group OpenID"
+    qq_private_openid = "QQ Private OpenID"
     qq_guild_channel = "QQ Guild Channel"
-    qq_guild_channel_open = "QQ Guild Channel Open"
     qq_guild_direct = "QQ Guild Direct"
-    qq_guild_direct_open = "QQ Guild Direct Open"
     kaiheila_channel = "Kaiheila Channel"
     kaiheila_private = "Kaiheila Private"
     unknown_ob12 = "Unknow Onebot 12 Platform"
@@ -30,6 +30,8 @@ class SupportedPlatform(StrEnum):
     telegram_forum = "Telegram Forum"
     feishu_private = "Feishu Private"
     feishu_group = "Feishu Group"
+    dodo_channel = "DoDo Channel"
+    dodo_private = "DoDo Private"
 
 
 supported_adapter_names = set(SupportedAdapters._member_map_.values())  # noqa: SLF001