+++ tests/conftest.py
@@ -3,6 +3,7 @@ import nonebot
 from nonebug import NONEBOT_INIT_KWARGS, App
 from nonebot.adapters.qq import Adapter as QQAdapter
 from nonebot.adapters.red import Adapter as RedAdapter
+from nonebot.adapters.dodo import Adapter as DoDoAdapter
 from nonebot.adapters.feishu import Adapter as FeishuAdapter
 from nonebot.adapters.qqguild import Adapter as QQGuildAdapter
 from nonebot.adapters.telegram import Adapter as TelegramAdapter
@@ -33,4 +34,5 @@ def load_adapters(nonebug_init: None):
     driver.register_adapter(TelegramAdapter)
     driver.register_adapter(FeishuAdapter)
     driver.register_adapter(RedAdapter)
+    driver.register_adapter(DoDoAdapter)
     driver.register_adapter(QQAdapter)