+++ tests/test_onebot_v11.py
@@ -235,13 +235,18 @@ async def test_send_aggreted_ob11(app: App):
 
 async def test_list_targets(app: App, mocker: MockerFixture):
     from nonebot_plugin_saa import TargetQQGroup, TargetQQPrivate
-    from nonebot_plugin_saa.auto_select_bot import get_bot, refresh_bots
+    from nonebot_plugin_saa.auto_select_bot import BOT_CACHE, get_bot, refresh_bots
 
     mocker.patch("nonebot_plugin_saa.auto_select_bot.inited", True)
 
     async with app.test_api() as ctx:
         adapter = get_adapter(Adapter)
         bot = ctx.create_bot(base=Bot, adapter=adapter)
+
+        ctx.should_call_api("get_group_list", {}, exception=Exception("test"))
+        ctx.should_call_api("get_friend_list", {}, exception=Exception("test"))
+        await refresh_bots()
+        assert not BOT_CACHE.get(bot)
 
         ctx.should_call_api("get_group_list", {}, [{"group_id": 112}])
         ctx.should_call_api("get_friend_list", {}, [{"user_id": 1122}])